Quietly exit PostWithdrawalJob if post not found

+++ app/jobs/post_withdrawal_job.rb
@@ -2,13 +2,12 @@ class PostWithdrawalJob < ApplicationJob
   queue_as :default
 
   def perform(post_id)
-    post = Post.find(post_id)
-    client = post.feed.access_token.build_client
+    post = Post.find_by(id: post_id)
+    return unless post
 
+    client = post.feed.access_token.build_client
     client.delete_post(post.freefeed_post_id)
   rescue FreefeedClient::Error => e
     Rails.logger.error("Failed to withdraw post #{post_id} from FreeFeed: #{e.message}")
-  rescue ActiveRecord::RecordNotFound
-    Rails.logger.error("Post #{post_id} not found")
   end
 end
